using variations for torch images
