+++ release/release_tests.yaml
@@ -118,25 +118,25 @@
 
   cluster:
     anyscale_sdk_2026: true
-    byod:
-      type: torch-cpu
     cluster_compute: hello_world_compute_config.yaml
 
   run:
     timeout: 1800
-    script: python hello_world_torch_cpu.py
+    script: python hello_world_torch.py
 
   variations:
     - __suffix__: cpu
+      cluster:
+        anyscale_sdk_2026: true
+        byod:
+          type: torch-cpu
+        cluster_compute: hello_world_compute_config.yaml
     - __suffix__: cu130
       cluster:
         anyscale_sdk_2026: true
         byod:
           type: torch-cu130
         cluster_compute: hello_world_compute_config_gpu.yaml
-      run:
-        timeout: 1800
-        script: python hello_world_torch_gpu.py
 
 - name: hello_world_custom_byod
   python: "3.10"
